# Simplify and optimize constant alias tables