@LilithHafner 2024-04-23
RegressionTests.jl is complaining about an allocation about regression in `@b rand(1) AliasTable` which is expected and nbd, and more than compensated for imo by the larger constants.