@LilithHafner 2024-04-24
This also fixes

```julia
julia> rand(AliasTable{UInt8}(vcat(fill(0, 500), 1)))
153

julia> extrema(rand(AliasTable{UInt8}(vcat(fill(0, 500), 1))) for _ in 1:1000)
(1, 255)
```